リファクタリング　一部不要になった処理があったので整理
