+++ .env.example
@@ -1,7 +1,7 @@
 # =============================================
 # アプリケーション環境変数設定例
 # =============================================
-# 使用方法: このファイルを「.env」にコピーして、値を環境に合わせて設定してください
+# 使用方法: このファイルを「.env」に変更して、値を環境に合わせて設定してください
 # =============================================
 
 # ================== データベース設定 ==================
@@ -22,13 +22,4 @@ API_PORT=8000
 # デバッグモード設定 (開発環境: true, 本番環境: false)
 DEBUG=false
 
-# キャッシュ有効期間 (秒単位、デフォルト: 3600秒 = 1時間)
-CACHE_TTL=3600
-
-# ================== パフォーマンス設定 ==================
-
-# テキスト処理の最大ワーカー数
-MAX_WORKERS=5
-
-# リクエスト制限 (分あたりの最大リクエスト数)
-RATE_LIMIT=100
+#
